# wasmi_v1: improve instruction scheduler