--- a/wasmi_v1/src/engine/bytecode/mod.rs
+++ b/wasmi_v1/src/engine/bytecode/mod.rs
@@ -1,15 +1,11 @@
 //! The instruction architecture of the `wasmi` interpreter.
 
 mod utils;
-mod visitor;
 
 #[cfg(test)]
 mod tests;
 
-pub use self::{
-    utils::{BrTable, DropKeep, FuncIdx, GlobalIdx, LocalIdx, Offset, SignatureIdx, Target},
-    visitor::VisitInstruction,
-};
+pub use self::utils::{DropKeep, FuncIdx, GlobalIdx, LocalIdx, Offset, SignatureIdx, Target};
 use wasmi_core::UntypedValue;
 
 /// The internal `wasmi` bytecode that is stored for Wasm functions.