--- a/wasmi_v1/src/engine/exec_context.rs
+++ b/wasmi_v1/src/engine/exec_context.rs
@@ -73,22 +73,22 @@ impl<'engine, 'func> ExecutionContext<'engine, 'func> {
         mut ctx: impl AsContextMut,
     ) -> Result<FunctionExecutionOutcome, Trap> {
         'outer: loop {
-            let pc = self.frame.inst_ptr;
+            let pc = self.frame.pc;
             let inst_context =
                 InstructionExecutionContext::new(self.value_stack, self.frame, &mut ctx);
             match self.func_body.visit(pc, inst_context)? {
                 ExecutionOutcome::Continue => {
                     // Advance instruction pointer.
-                    self.frame.inst_ptr += 1;
+                    self.frame.pc += 1;
                 }
                 ExecutionOutcome::Branch(target) => {
                     self.value_stack.drop_keep(target.drop_keep());
                     // Set instruction pointer to the branch target.
-                    self.frame.inst_ptr = target.destination_pc().into_usize();
+                    self.frame.pc = target.destination_pc().into_usize();
                 }
                 ExecutionOutcome::ExecuteCall(func) => {
                     // Advance instruction pointer.
-                    self.frame.inst_ptr += 1;
+                    self.frame.pc += 1;
                     return Ok(FunctionExecutionOutcome::NestedCall(func));
                 }
                 ExecutionOutcome::Return(drop_keep) => {