--- a/wasmi_v1/src/engine/mod.rs
+++ b/wasmi_v1/src/engine/mod.rs
@@ -376,44 +376,6 @@ impl EngineInner {
         &self.config
     }
 
-    // /// Unpacks the entity and checks if it is owned by the engine.
-    // ///
-    // /// # Panics
-    // ///
-    // /// If the guarded entity is not owned by the engine.
-    // fn unwrap_index<Idx>(&self, stored: Guarded<Idx>) -> Idx
-    // where
-    //     Idx: Index,
-    // {
-    //     stored.entity_index(self.engine_idx).unwrap_or_else(|| {
-    //         panic!(
-    //             "encountered foreign entity in engine: {}",
-    //             self.engine_idx.into_usize()
-    //         )
-    //     })
-    // }
-
-    // /// Allocates a new function type to the engine.
-    // pub(super) fn alloc_func_type(&mut self, func_type: FuncType) -> Signature {
-    //     Signature::from_inner(Guarded::new(
-    //         self.engine_idx,
-    //         self.func_types.alloc(func_type),
-    //     ))
-    // }
-
-    // /// Resolves a deduplicated function type into a [`FuncType`] entity.
-    // ///
-    // /// # Panics
-    // ///
-    // /// - If the deduplicated function type is not owned by the engine.
-    // /// - If the deduplicated function type cannot be resolved to its entity.
-    // pub(super) fn resolve_func_type(&self, func_type: Signature) -> &FuncType {
-    //     let entity_index = self.unwrap_index(func_type.into_inner());
-    //     self.func_types
-    //         .get(entity_index)
-    //         .unwrap_or_else(|| panic!("failed to resolve stored function type: {:?}", entity_index))
-    // }
-
     /// Allocates the instructions of a Wasm function body to the [`Engine`].
     ///
     /// Returns a [`FuncBody`] reference to the allocated function body.