This PR changes the instruction scheduler of the `wasmi_v1` engine.
The new approach seems promising and also is less affected by the performance regressions with newer `rustc` versions than the current interpreter instruction scheduler.
Note that the instruction scheduler is on the absolute hot path of the interpreter which is why we see significant changes in performance.

We compare the performance of Rust 1.59 because that is the last version before the performance regressions and we also use Rust 1.62 because it is the current stable Rust version.

## Rust 1.59

| Test | Before | After | Change |
|------|--------|-------|--------|
| `tiny_keccak` | 970.77 | 793.95 µs | <span style="color:green">-17.890%</span> |
| `rev_complement` | 1.1200 | 991.54 µs | <span style="color:green">-11.326%</span> |
| `regex_redux` | 1.1962 | 1.0809 ms | <span style="color:green">-9.5504%</span> |
| `count_until` | 1.7180 | 1.4201 ms | <span style="color:green">-17.382%</span> |
| `factorial_recursive` | 1.1851 | 1.1655 µs | <span style="color:green">-1.6003%</span> |
| `factorial_optimized` | 740.47 | 635.99 ns | <span style="color:green">-13.924%</span> |
| `recursive_ok` | 312.90 | 304.17 µs | <span style="color:green">-3.2702%</span> |
| `recursive_trap` | 28.212 | 27.951 µs | -0.8383% |
| `host_calls` | 45.054 | 43.287 µs | <span style="color:green">-3.7734%</span> |
| `fib_recursive` | 12.633 | 13.025 ms | <span style="color:red">+3.0991%</span> |

## Rust 1.62 (Stable)

| Test | Before | After | Change |
|------|--------|-------|--------|
| `tiny_keccak` | 1.4818 ms | 916.57 µs | <span style="color:green">-36.988%</span> |
| `rev_complement` | 1.4554 ms | 1.1097 ms | <span style="color:green">-23.228%</span> |
| `regex_redux` | 1.4766 ms | 1.1896 ms | <span style="color:green">-19.362%</span> |
| `count_until` | 2.2027 ms | 1.6776 ms | <span style="color:green">-23.720%</span> |
| `factorial_recursive` | 1.4349 µs | 1.2398 µs | <span style="color:green">-14.031%</span> |
| `factorial_optimized` | 935.66 ns | 719.33 ns | <span style="color:green">-23.144%</span> |
| `recursive_ok` | 354.32 µs | 322.68 µs | <span style="color:green">-8.6313%</span> |
| `recursive_trap` | 33.905 µs | 29.536 µs | <span style="color:green">-12.983%</span> |
| `host_calls` | 53.669 µs | 49.136 µs | <span style="color:green">-8.5481%</span> |
| `fib_recursive` | 15.320 ms | 13.496 ms | <span style="color:green">-11.907%</span> |

## Conclusion

All in all a significant improvement across the board. However, even the new instruction scheduler still suffers a performance degradation of roughly 15-20% with a newer `rustc` version compared to Rust 1.59.