@Robbepop 2022-07-13
Note that this PR also fixes [this `rustc` issue](https://github.com/rust-lang/rust/issues/95409) in parts since the new instruction dispatcher is no longer as heavily affected by the `rustc` performance regressions as was the case with the previous (old) instruction dispatcher. Between Rust 1.59 and Rust 1.62 (current stable) the old instruction dispatcher regressed by roughly 50-60% in performance according to our benchmarks whereas the new instruction dispatcher is "only" regressed by roughly 20%. Still 20% is a lot though ...